More script organization
