+++ DungeonManager/Assets/Scripts/Characters/Move_Main.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a40dbee71b0b4a2469d3fa1fd7112a2f
+guid: 728751f3ba0b0964892b4d4056e0bf4a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
