+++ DungeonManager/Assets/Scripts/Lighting/LightColliderOnStart.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7a7cd560bc46ef54cbc014ee4bdd9438
+guid: 45699d1b8de864142af3c39fa90bcdc7
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
