--- DungeonManager/Assets/Sound and Music.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d1ffe50c9713b2b48991827de9ac84e5
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
